New enumeration for buffer primitive types
LWJGL only allows certain types of data to be loaded into OpenGL
buffers.  These take the form of different methods to require different
types of Java Buffers.  This enumeration abstracts each of those
specific methods to an enumeration value.

It's sort of a hack and should be discussed some more.  Later.