Removed dependency on Game:30 parent pom file
The Game:30 parent pom file is not available to anyone outside of
Game:30 Studios.  This makes importing the JavaGL project for anyone
outside of Game:30 Studios impossible.  By removing the dependency,
others outside of Game:30 Studios will be able to work on, improve, and
contribute to JavaGL.

The dependency on the Game:30 parent pom file has not been removed as
it will still be used when building the project by the Game:30 Jenkins
server.